requirements.txt
gdal==3.2.1
affine==2.3.0
attrs==20.3.0
autopep8==1.5.4
cdsapi==0.5.1
certifi==2020.11.8
cftime==1.3.0
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
colorama==0.4.4
future==0.18.2
geojson==2.5.0
geomet==0.3.0
geoserver-restconfig==2.0.4.8
gisdata==0.5.4
html2text==2020.1.16
idna==2.10
importlib-metadata==4.8.1
netCDF4==1.5.4
numpy==1.19.3
pandas==0.25.3
psycopg2==2.8.6
pycodestyle==2.6.0
pyparsing==2.4.7
pyproj==3.0.0.post1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
rioxarray==0.3.1
scipy==1.5.4
sentinelsat==1.0.1
Shapely==1.7.1
six==1.15.0
slug==2.0
snuggs==1.4.7
toml==0.10.2
tqdm==4.53.0
typing-extensions==3.10.0.2
urllib3==1.26.2
xarray==0.16.1
zipp==3.5.0
